pyproject.toml
[project]
name = "fashion-designing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "imagehash>=4.3.2",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "tqdm>=4.67.1",
    "ultralytics>=8.3.216",
]
